+++ server/routes/genesetEnrichment.ts
@@ -78,7 +78,7 @@ function init({ genomes }) {
 				}
 			})
 		} catch (e: any) {
-			res.send({ status: 'error', error: e.message || e })
+			res.status(e.status || 500).send({ status: 'error', error: e.message || e, code: e.code })
 			if (e.stack) console.log(e.stack)
 		}
 	}